# openapi.yaml
openapi: 3.1.0
info:
  title: SecGuard API
  version: 0.2.0
  description: API for application discovery, model analysis, repository scans, and stored findings.
servers:
  - url: /
paths:
...
  /api/repository/scans:
    post:
      summary: Store a repository scanner report
      description: Accepts the JSON report emitted by `python -m backend.repository scan`.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
  /api/applications/{application_id}/sbom:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      summary: Download a repository software bill of materials
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [cyclonedx, spdx]
            default: cyclonedx
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/applications/{application_id}/quality:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      summary: Get repository quality metrics and health
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
  /api/repositories:
    get:
      summary: List repositories available to the current user
      security:
        - cookieAuth: []
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
  /api/repositories/refresh:
    post:
      summary: Refresh the signed-in user's repository list
      security:
        - cookieAuth: []
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/repositories/{owner}/{name}/scan:
    parameters:
      - in: path
        name: owner
        required: true
        schema: { type: string }
      - in: path
        name: name
        required: true
        schema: { type: string }
    post:
      summary: Trigger the SecGuard workflow for a repository
      security:
        - cookieAuth: []
      responses:
        "202": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/repositories/{owner}/{name}/name:
    parameters:
      - in: path
        name: owner
        required: true
        schema: { type: string }
      - in: path
        name: name
        required: true
        schema: { type: string }
    put:
      summary: Rename a repository application
      security:
        - cookieAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Required when SECGUARD_API_TOKEN is configured.
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: GitHub sign-in session cookie.
  parameters:
    ApplicationId:
      in: path
      name: application_id
      required: true
      schema: { type: string }
  schemas:
...
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
    Success:
      type: object
      properties:
        success: { type: boolean, example: true }
      additionalProperties: true
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Success:
      description: Request completed successfully.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Success" }